修复很多不合理的地方

+++ src/main/java/run/app/translator/service/TranslateService.java
@@ -82,12 +82,16 @@ public class TranslateService {
 
     public void deleteByStrings(Strings strings) {
         Iterable<Translate> translates = findAllByStrings(strings);
-        translateRepository.deleteInBatch(translates);
+        if (translates.iterator().hasNext()) {
+            translateRepository.deleteInBatch(translates);
+        }
     }
 
     public void deleteByStringsId(Integer stringsId) {
         Iterable<Translate> translates = findTranslateByStrings(stringsId);
-        translateRepository.deleteInBatch(translates);
+        if (translates.iterator().hasNext()) {
+            translateRepository.deleteInBatch(translates);
+        }
     }
 
     public void delete(Integer id) {
